+++ betterwork.backend/betterwork.Data/src/main/java/com/erosion/entity/Employee.java
@@ -1,4 +1,4 @@
-package com.erosion.models;
+package com.erosion.entity;
 
 import javax.persistence.*;
 import java.io.Serializable;
