+++ betterwork.backend/betterwork.Data/src/main/java/com/erosion/entity/Post.java
@@ -1,4 +1,4 @@
-package com.erosion.models;
+package com.erosion.entity;
 
 
 import javax.persistence.*;
